Create L. Snack
Given a number X which donates a code and a real number Y which donates the amount of an item. According to the following table, Print the total price of buying Y amount of the product that has X code.


Input
Only one line containing X, Y (1 ≤ X ≤ 5, 1.0 ≤ Y ≤ 100.0)

Output
Print "Total: R " followed by the answer to the problem above, with 2 digits after the decimal point.